+++ tests/benchmark/Bench/S50kBench.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace KSamuel\FacetedSearch\Tests\Benchmark\Bench;
+
+use KSamuel\FacetedSearch\Filter\FilterInterface;
+use KSamuel\FacetedSearch\Index;
+use KSamuel\FacetedSearch\Search;
+use KSamuel\FacetedSearch\Filter\ValueFilter;
+use KSamuel\FacetedSearch\Tests\Benchmark\DatasetFactory;
+use KSamuel\FacetedSearch\Tests\Benchmark\SearchBench;
+use PhpBench\Benchmark\Metadata\Annotations\BeforeMethods;
+use PhpBench\Benchmark\Metadata\Annotations\Iterations;
+use PhpBench\Benchmark\Metadata\Annotations\Revs;
+
+/**
+ * @Iterations(5)
+ * @Revs(10)
+ * @BeforeMethods({"before"})
+ */
+class S50KBench extends SearchBench
+{
+    /**
+     * @var int
+     */
+    protected $dataSize = 50000;
+}